Add more tests on gurobi_torch_mip objective gradient. 1. round the binary variable value to the nearest integer, gurobi sometimes return solution that are not exactly 0 or 1. 2. Replace the for loop in gurobi_torch_mip::get_active_constraints with constructing torch sparse matrix. This is faster.

+++ robust_value_approx/lyapunov.py
@@ -60,8 +60,8 @@ class LyapunovHybridLinearSystem:
 
         # Now add the constraint that sum gamma = 1
         milp.addLConstr(
-            [torch.ones((self.system.num_modes,))], [gamma],
-            sense=gurobipy.GRB.EQUAL, rhs=1.)
+            [torch.ones((self.system.num_modes,), dtype=self.system.dtype)],
+            [gamma], sense=gurobipy.GRB.EQUAL, rhs=1.)
         return (x, s, gamma, Aeq_s, Aeq_gamma)
 
     def add_relu_output_constraint(
